refactor: move shared policy types to bitrouter-core
Move PolicyConfig, PolicyFile, PolicyContext, PolicyResult,
ToolProviderPolicy, and policy_dir() from bitrouter-accounts to
bitrouter-core::policy. These types have no account-specific
dependencies and are used across crates.

bitrouter-accounts::policy::file retains load_policies() (I/O +
tracing) and re-exports the core types for backwards compatibility.
