+++ bitrouter-core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod errors;
 pub mod hooks;
 pub mod models;
 pub mod observe;
+pub mod policy;
 pub mod pricing;
 pub mod routers;
 pub mod sync;
